move tests under graph
